+++ SimulationControl.cpp
@@ -249,7 +249,7 @@ bool LoadSimulation(Dataport *loader) {
 	for (i = 0; i < sHandle->totalFacet; i++) {
 		FacetProperties *shFacet = (FacetProperties *)buffer;
 		sHandle->str[shFacet->superIdx].nbFacet++;
-		buffer += sizeof(FacetProperties) + shFacet->nbIndex*(sizeof(int) + sizeof(Vector2d));
+		buffer += sizeof(FacetProperties) + shFacet->nbIndex*(sizeof(size_t) + sizeof(Vector2d));
 		if (shFacet->useOutgassingFile) buffer += sizeof(double)*shFacet->outgassingMapWidth*shFacet->outgassingMapHeight;
 		if (shFacet->anglemapParams.hasRecorded) buffer += sizeof(size_t)*shFacet->anglemapParams.phiWidth*(shFacet->anglemapParams.thetaLowerRes+shFacet->anglemapParams.thetaHigherRes);
 		if (shFacet->isTextured) buffer += sizeof(double)*shFacet->texWidth*shFacet->texHeight;
@@ -336,10 +336,10 @@ bool LoadSimulation(Dataport *loader) {
 
 		// Reset counter in local memory
 		//memset(&(f->sh.counter), 0, sizeof(FacetHitBuffer));
-		f->indices = (int *)malloc(f->sh.nbIndex * sizeof(int));
+		f->indices = (size_t *)malloc(f->sh.nbIndex * sizeof(size_t));
 		buffer += sizeof(FacetProperties);
-		memcpy(f->indices, buffer, f->sh.nbIndex * sizeof(int));
-		buffer += f->sh.nbIndex * sizeof(int);
+		memcpy(f->indices, buffer, f->sh.nbIndex * sizeof(size_t));
+		buffer += f->sh.nbIndex * sizeof(size_t);
 		f->vertices2 = (Vector2d *)malloc(f->sh.nbIndex * sizeof(Vector2d));
 		if (!f->vertices2) {
 			SetErrorSub("Not enough memory to load vertices");
